requirements.txt
discord.py
dotenv.py
os.py
requests.py

